requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
psycopg2-binary==2.9.9
sentence-transformers==2.3.1
requests==2.31.0
torch==2.1.2
transformers==4.36.2
numpy==1.24.3